// src/types.ts
export type NotificationSound = 
  | 'chime_modern' 
  | 'marimba_bell' 
  | 'futuristic_ding' 
  | 'subtle_pop' 
  | 'urgent_alert' 
  | 'elevator_ding' 
  | 'crystal_shine' 
  | 'radar_ping'
  | 'retro_arcade'
  | 'zen_bowl'
  | 'harpa_celeste'
  | 'cash_register'
  | 'whistle_bright'
  | 'cyber_synth'
  | 'kalimba_dream'
  | 'magic_sparkle';